{{ cookiecutter.project_slug }}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.project_slug }}"
version = "0.1.0"
description = ""
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
coderedcms = "^4.0.1"
django-storages = "^1.14.4"
boto3 = "^1.34.157"
django-redis = "^5.4.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^22.0.0"
sentry-sdk = "^2.12.0"
whitenoise = "^6.7.0"
django-htmx = "^1.19.0"
django-render-block = "^0.10"
django-environ = "^0.11.2"
django-dbbackup = "^4.1.0"
libsass = "^0.23.0"
django-compressor = "^4.5.1"
django-sass-processor = "^1.4.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
django-coverage-plugin = "^3.1.0"
django-debug-toolbar = "^4.4.6"
django-stubs = "^5.0.4"
djlint = "^1.34.1"
flake8 = "^7.1.1"
flake8-isort = "^6.1.1"
ipdb = "^0.13.13"
mypy = "^1.11.1"
pre-commit = "^3.8.0"
pylint-django = "^2.5.5"
pytest = "^8.3.2"
pytest-django = "^4.8.0"
pytest-sugar = "^1.0.0"
coverage = "^7.6.1"

[tool.black]
line-length = 80
extend-exclude = ["migrations"]

[tool.django-stubs]
django_settings_module = "{{ cookiecutter.project_slug }}.settings"

[tool.mypy]
ignore_missing_imports = true
plugins = ["mypy_django_plugin.main"]
exclude = [
    '^\..*',
    'migrations',
    'node_modules',
    'venv',
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "{{ cookiecutter.project_slug }}.settings"
addopts = "--cov --cov-report html"
python_files = "tests.py test_*.py"

[tool.ruff]
extend-exclude = ["migrations"]
line-length = 80

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort]
case-sensitive = false
force-single-line = true
lines-after-imports = 2

[tool.djlint]
use_gitignore=true
ignore="H006,H013,H016"
